[fix] compilation warning

+++ src/lsize.c
@@ -10,7 +10,7 @@
 inline size_t lbuffer_lsize(lbuffer_t *buffer)
 {
   if (buffer->reader < buffer->writer)
-    return (buffer->size - buffer->writer + buffer->reader);
+    return (buffer->size - (size_t)buffer->writer + (size_t)buffer->reader);
   else if (buffer->reader > buffer->writer)
     return (buffer->reader - buffer->writer);
   return ((buffer->empty) ? buffer->size : 0);
